test(product): split BackgroundWorkPane.test.tsx for PROD-SIZE-1
R2-R5 review rounds accumulated coverage onto the one file until it hit
690 lines against the repo-wide 600-line cap (scripts/check_max_lines.py).
Mechanical split, no product code touched and no assertions
added/removed/weakened: roster/scopes/seam/detail-routing tests stay in
BackgroundWorkPane.test.tsx (12 tests); the R5-era finish-signal ladder
concerns (NoticeBanner, pending-selection consumption, feed
isOpen-gating) move verbatim to the new sibling
BackgroundWorkPane.finish-signals.test.tsx (12 tests). Shared mock
components and value builders extracted to
BackgroundWorkPane.test.fixtures.tsx (145 lines) so neither test file
re-declares them.

The fixtures file is named with a literal ".test." segment (not
".test-fixtures") so it is recognized as test-only scaffolding by both
report_frontend_structure.py's should_skip (exempting its mock JSX from
the shipped-component raw-DOM-control rule) and vitest's *.test.tsx
include glob (it is not itself collected as a test suite).

24 tests total before and after (was 1 file/24 tests, now 2 files/12+12).
